TreeWidthSolver

Documentation for TreeWidthSolver.

TreeWidthSolver.MaskedBitGraph

TreeWidthSolver.elimination_order

TreeWidthSolver.exact_treewidth

TreeWidthSolver.graph_from_gr

TreeWidthSolver.line_graph

TreeWidthSolver.save_graph

TreeWidthSolver.simple_graph

TreeWidthSolver.MaskedBitGraph — Type

mutable struct MaskedBitGraph{INT}

A mutable struct representing a masked bit graph.

Fields

MaskedBitGraph::Vector{INT}: Stores the adjacency matrix as a vector of BitStr.

fadjlist::Vector{Vector{Int}}: Stores the adjacency list, providing information about the sparse graph.

mask::INT: The mask for the graph.

TreeWidthSolver.elimination_order — Method

elimination_order(g::SimpleGraph{TG}; labels::Vector{TL} = collect(1:nv(g)), weights::Vector{TW} = ones(nv(g)), verbose::Bool = false) where {TG, TL, TW}

Compute the elimination order of a graph g using the BT algorithm.

Arguments

g::SimpleGraph{TG}: The input graph.

labels::Vector{TL}: (optional) Labels for the vertices of g. Default is collect(1:nv(g)).

weights::Vector{TW}: (optional) Weights for the vertices of g. Default is ones(nv(g)).

verbose::Bool: (optional) Whether to print verbose output. Default is false.

Returns

labeled_eo::Vector{Vector{TL}}: The elimination order of the graph g, where each vertex is labeled according to labels.

TreeWidthSolver.exact_treewidth — Method

exact_treewidth(g::SimpleGraph{TG}; weights::Vector{TW} = ones(nv(g)), verbose::Bool = false) where {TG, TW}

Compute the exact treewidth of a given graph g using the BT algorithm.

Arguments

g::SimpleGraph{TG}: The input graph.

weights::Vector{TW} = ones(nv(g)): The weights of the vertices in the graph. Default is equal weights for all vertices.

verbose::Bool = false: Whether to print verbose output. Default is false.

Returns

tw: The treewidth of the graph.

TreeWidthSolver.graph_from_gr — Method

graph_from_gr(filename)

Reads a graph from a file in the .gr format (PACE format) and returns a SimpleGraph object.

Arguments

filename: The path to the input file.

Returns

graph: A SimpleGraph object representing the graph.

TreeWidthSolver.line_graph — Method

line_graph(adjacency_mat::SparseMatrixCSC)

Constructs the line graph of a given graph represented by its adjacency matrix.

Arguments

adjacency_mat::SparseMatrixCSC: The adjacency matrix of the input graph, where the columns represent the vertices and the rows represent the edges. The value is true if the edge is connected to the vertex.

Returns

g::SimpleGraph: The line graph of the input graph.

TreeWidthSolver.save_graph — Method

function save_graph(g::SimpleGraph, filename)

The graph will be saved as .gr format, in PACE format, where the first line is p tw nv ne, and the following lines are the edges src dst

TreeWidthSolver.simple_graph — Method

simple_graph(adjacency_mat::SparseMatrixCSC)

Constructs a simple undirected graph from a sparse adjacency matrix.

Arguments

adjacency_mat::SparseMatrixCSC: The sparse adjacency matrix representing the graph.

Returns

g::SimpleGraph: The constructed simple undirected graph.